Various small fixes to avoid warnings with gcc-4.0 and gcc-4.1. It would have be nice to commit this in smaller pieces, but it would require too much work.
Patch from Markus F.X.J. Oberhumer <markus@oberhumer.com>
(http://www.delorie.com/djgpp/mail-archives/browse.cgi?p=djgpp-workers/2006/06/12/06:28:05)

+++ src/libc/dos/lfn/_use_lfn.c
@@ -159,10 +159,10 @@ _use_lfn (const char *path)
       same_drive_as_last_time = 0;
     else
     {
-      int drv_no;
+      unsigned drv_no;
 
       _dos_getdrive(&drv_no);
-      if (drv_no - 1 + 'A' != last_drive)
+      if ((int)drv_no - 1 + 'A' != last_drive)
 	same_drive_as_last_time = 0;
     }
   }
